Check settings.inc.php is writeable and remove junk
Added checks to see whether /inc/settings.inc.php is actually writeable
and catch the output of the actual write operation to check if writing
the settings file went OK. This resolves #7.

Also I detected a not working piece of code (on my box) that was
supposed to display the modified time of the install script. Don’t know
why anybody would care, so I trashed it.